refactor(ai): split dashscope json parsing
